+++ src/constants/data.ts
@@ -1,0 +1,33 @@
+export const brokers = [
+  {
+    id: "1",
+    name: "Zerodha",
+    logo: "https://assets.smallcase.com/smallcase/assets/brokerLogo/small/kite.svg",
+  },
+  {
+    id: "2",
+    name: "Groww",
+    logo: "https://assets.smallcase.com/smallcase/assets/brokerLogo/small/groww.svg",
+  },
+  {
+    id: "3",
+    name: "Upstocks",
+    logo: "https://assets.smallcase.com/smallcase/assets/brokerLogo/small/upstox.svg",
+  },
+  {
+    id: "4",
+    name: "Angel One",
+    logo: "https://assets.smallcase.com/smallcase/assets/brokerLogo/small/angelbroking.svg",
+  },
+
+  {
+    id: "5",
+    name: "Kotak Sec",
+    logo: "https://assets.smallcase.com/smallcase/assets/brokerLogo/small/kotak.svg",
+  },
+  {
+    id: "6",
+    name: "HDFC Sec",
+    logo: "https://assets.smallcase.com/smallcase/assets/brokerLogo/small/hdfc.svg",
+  },
+];