+++ src/pages/Holdings/index.tsx
@@ -1,0 +1,5 @@
+const HoldingsPage = () => {
+  return <div>HoldingsPage</div>;
+};
+
+export default HoldingsPage;