+++ src/pages/Orderbook/index.tsx
@@ -1,0 +1,5 @@
+const OrderbookPage = () => {
+  return <div>OrderbookPage</div>;
+};
+
+export default OrderbookPage;