+++ src/pages/Positions/index.tsx
@@ -1,0 +1,5 @@
+const PositionsPage = () => {
+  return <div>PositionsPage</div>;
+};
+
+export default PositionsPage;